+++ src/controllers/grinder.controller.ts
@@ -21,6 +21,9 @@ export class GrinderController {
 
   async one(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        throw new Error('Invalid grinder ID. ID must be a string.');
+      }
       const grinder = await this.grinderService.getGrinderById(request.params.id);
       response.status(200).json(grinder);
     } catch (error) {
@@ -57,6 +60,10 @@ export class GrinderController {
     try {
       const { model, manufacturer, burrType, burrInstallDate, serialNumber } = request.body;
 
+      if (typeof request.params.id !== 'string') {
+        throw new Error('Invalid grinder ID. ID must be a string.');
+      }
+
       const result = await this.grinderService.updateGrinder(request.params.id, {
         model,
         manufacturer,
@@ -79,6 +86,9 @@ export class GrinderController {
 
   async remove(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        throw new Error('Invalid grinder ID. ID must be a string.');
+      }
       await this.grinderService.deleteGrinder(request.params.id);
       response.status(204).send();
     } catch (error) {
