+++ src/controllers/shotEnvironment.controller.ts
@@ -23,6 +23,9 @@ export class ShotEnvironmentController {
 
   async one(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        throw new Error('Invalid shot environment ID. ID must be a string.');
+      }
       const environment = await this.shotEnvironmentService.getShotEnvironmentById(
         request.params.id
       );
@@ -40,6 +43,9 @@ export class ShotEnvironmentController {
 
   async save(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        throw new Error('Invalid shot ID. ID must be a string.');
+      }
       const {
         ambient_temp_c,
         humidity_percent,
@@ -70,6 +76,9 @@ export class ShotEnvironmentController {
 
   async update(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        throw new Error('Invalid shot environment ID. ID must be a string.');
+      }
       const {
         ambient_temp_c,
         humidity_percent,
@@ -99,6 +108,9 @@ export class ShotEnvironmentController {
 
   async remove(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        throw new Error('Invalid shot environment ID. ID must be a string.');
+      }
       await this.shotEnvironmentService.deleteShotEnvironment(request.params.id);
       response.status(204).send();
     } catch (error) {
