+++ src/controllers/shotFeedback.controller.ts
@@ -21,6 +21,9 @@ export class ShotFeedbackController {
 
   async one(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        throw new Error('Invalid shot feedback ID. ID must be a string.');
+      }
       const feedback = await this.shotFeedbackService.getShotFeedbackById(request.params.id);
 
       if (!feedback) {
@@ -39,6 +42,9 @@ export class ShotFeedbackController {
 
   async save(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        throw new Error('Invalid shot ID. ID must be a string.');
+      }
       const { overall_score, acidity, sweetness, bitterness, body, extraction_assessment, notes } =
         request.body;
 
@@ -64,6 +70,9 @@ export class ShotFeedbackController {
 
   async update(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        throw new Error('Invalid shot feedback ID. ID must be a string.');
+      }
       const { overall_score, acidity, sweetness, bitterness, body, extraction_assessment, notes } =
         request.body;
 
@@ -88,6 +97,9 @@ export class ShotFeedbackController {
 
   async remove(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        throw new Error('Invalid shot feedback ID. ID must be a string.');
+      }
       await this.shotFeedbackService.deleteShotFeedback(request.params.id);
       response.status(204).send();
     } catch (error) {
