+++ src/controllers/shotPreparation.controller.ts
@@ -24,6 +24,9 @@ export class ShotPreparationController {
 
   async one(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        throw new Error('Invalid shot preparation ID. ID must be a string.');
+      }
       const preparation = await this.shotPreparationService.getShotPreparationById(
         request.params.id
       );
@@ -41,6 +44,9 @@ export class ShotPreparationController {
 
   async save(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        throw new Error('Invalid shot ID. ID must be a string.');
+      }
       const {
         shot_id,
         dose_grams,
@@ -74,6 +80,9 @@ export class ShotPreparationController {
 
   async update(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        throw new Error('Invalid shot preparation ID. ID must be a string.');
+      }
       const {
         dose_grams,
         grind_setting,
@@ -105,6 +114,9 @@ export class ShotPreparationController {
 
   async remove(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        throw new Error('Invalid shot preparation ID. ID must be a string.');
+      }
       await this.shotPreparationService.deleteShotPreparation(request.params.id);
 
       response.status(204).send();
